refactor(chatroom): move stream-agent under session-group seam (phase 2)
Relocate streamAgentTurn + sendToAgentWithRetry + TurnTimeoutError from packages/services/src/chatroom/orchestration/stream-agent.ts to packages/services/src/session-group/stream-session.ts. The old path is now a thin re-export shim so existing strategies (Concurrent, Sequential, GroupChat, Handoff, Magentic) keep compiling without churn.

Move the colocated tests alongside the new home. No behavior change; all 1105 vitest specs pass.

Phase 5 deletes the shim once strategies relocate under session-group/orchestrators/.
