+++ packages/services/src/session-group/index.ts
@@ -1,3 +1,14 @@
 export { SessionGroup } from './SessionGroup';
 export { createApprovalGatePermissionFactory } from './permission-handler';
+export {
+  TurnTimeoutError,
+  streamAgentTurn,
+  sendToAgentWithRetry,
+} from './stream-session';
+export type {
+  StreamAgentOptions,
+  StreamAgentResult,
+  SendToAgentOptions,
+  SendToAgentResult,
+} from './stream-session';
 export type { PermissionHandlerFactory, SessionGroupSessionFactory } from './types';
